The Swiss Stock Exchange Market Data News via the SIX Group Newsboard

Returns the Swiss Stock Exchange Market Data News via the SIX Group Newsboard

Usage

newsboard_data(
  firstDate = Sys.Date() - 365 * 2,
  lastDate = Sys.Date(),
  messageType = c("Emergency Information", "Mistrade", "Suspension",
    "Trading Information"),
  Market = c("Six Swiss Exchange", "Structured Products",
    "XBTR - Bilateral Trading Plattform"),
  Products = c("Bonds", "Derivatives", "Equities", "Exchange Traded Products", "Funds")
)

Arguments

firstDate

a character string indicating the year-month-date of From date. The default is equivalent to firstDate = Sys.Date() - 365*2, the earliest date for which data is available on the Newsboard.

lastDate

a character string indicating the year-month-date of To date. The default is equivalent to `lastDate = Sys.Date()``, the last date in range for which data is available, but not greater than current date on the Newsboard.

messageType

a vector of character strings of the message type. This must be one or combination of either of the strings "Emergency Information", "Mistrade", "Suspension", "Trading Information".The default is equivalent to messageType = 'All'.

Market

a vector of character strings of the Market type. This must be one or combination of either of the strings "Six Swiss Exchange", "Structured Products","XBTR - Bilateral Trading Plattform". The default is equivalent to Market = 'All'.

Products

a vector of character strings of the products type. This must be one or combination of either of the strings "Bonds", "Derivatives", "Equities", "Exchange Traded Products", "Funds".The default is equivalent to Products = 'All'.

Value

Swiss Stock Exchange Market Data News via the SIX Group Newsboard

Details

The inputs must all be character strings, with the date format as %Y-%m-%d (as determined by as.Date()). The 'messageType', 'Market', 'Products', if not supplied, all character string options are used, which leads to the messageType = 'All', Market = 'All' and Products = 'All' defaults. Supplying individual parameter requires the use of complete string as opposed to the use of abbreviations.

References

Newsboard: Market activity at the Swiss Stock Exchange. SIX. https://www.six-group.com/en/products-services/the-swiss-stock-exchange/market-data/news-tools/newsboard.html#/

Author

Akhapwoya Obuchere

Examples

library(plyr)
library(jsonlite)

data_final <- newsboard_data(firstDate = '2023-06-24')
#> [1] "Page 0"
#> [1] "Page 0 has 10 observations."
#> [1] "Page 1"
#> [1] "Page 1 has 10 observations."
#> [1] "Page 2"
#> [1] "Page 2 has 10 observations."
#> [1] "Page 3"
#> [1] "Page 3 has 10 observations."
#> [1] "Page 4"
#> [1] "Page 4 has 10 observations."
#> [1] "Page 5"
#> [1] "Page 5 has 10 observations."
#> [1] "Page 6"
#> [1] "Page 6 has 10 observations."
#> [1] "Page 7"
#> [1] "Page 7 has 10 observations."
#> [1] "Page 8"
#> [1] "Page 8 has 10 observations."
#> [1] "Page 9"
#> [1] "Page 9 has 10 observations."
#> [1] "Page 10"
#> [1] "Page 10 has 10 observations."
#> [1] "Page 11"
#> [1] "Page 11 has 10 observations."
#> [1] "Page 12"
#> [1] "Page 12 has 10 observations."
#> [1] "Page 13"
#> [1] "Page 13 has 10 observations."
#> [1] "Page 14"
#> [1] "Page 14 has 10 observations."
#> [1] "Page 15"
#> [1] "Page 15 has 10 observations."
#> [1] "Page 16"
#> [1] "Page 16 has 10 observations."
#> [1] "Page 17"
#> [1] "Page 17 has 10 observations."
#> [1] "Page 18"
#> [1] "Page 18 has 10 observations."
#> [1] "Page 19"
#> [1] "Page 19 has 10 observations."
#> [1] "Page 20"
#> [1] "Page 20 has 10 observations."
#> [1] "Page 21"
#> [1] "Page 21 has 10 observations."
#> [1] "Page 22"
#> [1] "Page 22 has 10 observations."
#> [1] "Page 23"
#> [1] "Page 23 has 10 observations."
#> [1] "Page 24"
#> [1] "Page 24 has 10 observations."
#> [1] "Page 25"
#> [1] "Page 25 has 10 observations."
#> [1] "Page 26"
#> [1] "Page 26 has 10 observations."
#> [1] "Page 27"
#> [1] "Page 27 has 10 observations."
#> [1] "Page 28"
#> [1] "Page 28 has 10 observations."
#> [1] "Page 29"
#> [1] "Page 29 has 10 observations."
#> [1] "Page 30"
#> [1] "Page 30 has 2 observations."
#> [1] "Page 31"
#> [1] "Page 31 has  observations."
head(data_final, 1)
#>   messageNo         isin valorSymbol                                      title
#> 1    209039 CH1139756881      ZSMIAZ Mistrade Decision in ZSMIAZ / CH1139756881
#>   messageType broadcastDateTime     security      tradingSegment priority
#> 1    Mistrade       2.02402e+13 ZSMIAZ ZKB C Structured Products   Normal
#>   markets products currency
#> 1    XQMH       DE      CHF
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                          newsText
#> 1 <div><p>In accordance with the rules of SIX Swiss Exchange, the following trade in <strong>'ZSMIAZ ZKB C'</strong> has been declared a mistrade and has therefore been cancelled: </p><table><thead><tr><th align="center" style="width: 75px">Trade Date</th><th align="center" style="width: 75px">Time</th><th align="center" style="width: 45px">Cur</th><th align="left" style="width: 75px">Size</th><th align="left" style="width: 75px">Price</th><th align="left" style="width: 75px">Trade Type</th><th align="left" style="width: 75px">Book Type</th><th align="left" style="width: 75px">Ref Exch</th></tr></thead><tbody><tr><td align="center">02.02.2024</td><td align="center">09:15:58</td><td align="center">CHF</td><td align="left">17</td><td align="left">875.0000</td><td align="left">OnExchange</td><td align="left">QuoteBook</td><td align="left"/></tr></tbody></table><p>Please find further information concerning mistrades in Directive 4: Market Control on our website.</p><p>Regards,<br/>Exchange Operations, SIX Swiss Exchange</p></div>
#>   newsTypeCode
#> 1           MI
data_final <- newsboard_data(firstDate = '2023-06-24', messageType = 'Mistrade')
#> [1] "Page 0"
#> [1] "Page 0 has 10 observations."
#> [1] "Page 1"
#> [1] "Page 1 has 10 observations."
#> [1] "Page 2"
#> [1] "Page 2 has 10 observations."
#> [1] "Page 3"
#> [1] "Page 3 has 10 observations."
#> [1] "Page 4"
#> [1] "Page 4 has 10 observations."
#> [1] "Page 5"
#> [1] "Page 5 has 10 observations."
#> [1] "Page 6"
#> [1] "Page 6 has 10 observations."
#> [1] "Page 7"
#> [1] "Page 7 has 10 observations."
#> [1] "Page 8"
#> [1] "Page 8 has 10 observations."
#> [1] "Page 9"
#> [1] "Page 9 has 10 observations."
#> [1] "Page 10"
#> [1] "Page 10 has 10 observations."
#> [1] "Page 11"
#> [1] "Page 11 has 10 observations."
#> [1] "Page 12"
#> [1] "Page 12 has 10 observations."
#> [1] "Page 13"
#> [1] "Page 13 has 10 observations."
#> [1] "Page 14"
#> [1] "Page 14 has 10 observations."
#> [1] "Page 15"
#> [1] "Page 15 has 10 observations."
#> [1] "Page 16"
#> [1] "Page 16 has 10 observations."
#> [1] "Page 17"
#> [1] "Page 17 has 10 observations."
#> [1] "Page 18"
#> [1] "Page 18 has 10 observations."
#> [1] "Page 19"
#> [1] "Page 19 has 10 observations."
#> [1] "Page 20"
#> [1] "Page 20 has 10 observations."
#> [1] "Page 21"
#> [1] "Page 21 has 10 observations."
#> [1] "Page 22"
#> [1] "Page 22 has 10 observations."
#> [1] "Page 23"
#> [1] "Page 23 has 10 observations."
#> [1] "Page 24"
#> [1] "Page 24 has 10 observations."
#> [1] "Page 25"
#> [1] "Page 25 has 7 observations."
#> [1] "Page 26"
#> [1] "Page 26 has  observations."
head(data_final, 2)
#>   messageNo         isin valorSymbol                                      title
#> 1    209039 CH1139756881      ZSMIAZ Mistrade Decision in ZSMIAZ / CH1139756881
#> 2    209038 LU0950674761      EMUSRI Mistrade Decision in EMUSRI / LU0950674761
#>   messageType broadcastDateTime                    security      tradingSegment
#> 1    Mistrade       2.02402e+13                ZSMIAZ ZKB C Structured Products
#> 2    Mistrade       2.02402e+13 UBSETF MSCI EMU SRI EUR ACC                 ETF
#>   priority markets products currency
#> 1   Normal    XQMH       DE      CHF
#> 2   Normal    XSWX       FU      EUR
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                         newsText
#> 1                <div><p>In accordance with the rules of SIX Swiss Exchange, the following trade in <strong>'ZSMIAZ ZKB C'</strong> has been declared a mistrade and has therefore been cancelled: </p><table><thead><tr><th align="center" style="width: 75px">Trade Date</th><th align="center" style="width: 75px">Time</th><th align="center" style="width: 45px">Cur</th><th align="left" style="width: 75px">Size</th><th align="left" style="width: 75px">Price</th><th align="left" style="width: 75px">Trade Type</th><th align="left" style="width: 75px">Book Type</th><th align="left" style="width: 75px">Ref Exch</th></tr></thead><tbody><tr><td align="center">02.02.2024</td><td align="center">09:15:58</td><td align="center">CHF</td><td align="left">17</td><td align="left">875.0000</td><td align="left">OnExchange</td><td align="left">QuoteBook</td><td align="left"/></tr></tbody></table><p>Please find further information concerning mistrades in Directive 4: Market Control on our website.</p><p>Regards,<br/>Exchange Operations, SIX Swiss Exchange</p></div>
#> 2 <div><p>In accordance with the rules of SIX Swiss Exchange, the following trade in <strong>'UBSETF MSCI EMU SRI EUR ACC'</strong> has been declared a mistrade and has therefore been cancelled: </p><table><thead><tr><th align="center" style="width: 75px">Trade Date</th><th align="center" style="width: 75px">Time</th><th align="center" style="width: 45px">Cur</th><th align="left" style="width: 75px">Size</th><th align="left" style="width: 75px">Price</th><th align="left" style="width: 75px">Trade Type</th><th align="left" style="width: 75px">Book Type</th><th align="left" style="width: 75px">Ref Exch</th></tr></thead><tbody><tr><td align="center">01.02.2024</td><td align="center">16:00:29</td><td align="center">EUR</td><td align="left">200</td><td align="left">22.6600</td><td align="left">OnExchange</td><td align="left">QuoteBook</td><td align="left"/></tr></tbody></table><p>Please find further information concerning mistrades in Directive 4: Market Control on our website.</p><p>Regards,<br/>Exchange Operations, SIX Swiss Exchange</p></div>
#>   newsTypeCode
#> 1           MI
#> 2           MI
data_final <- newsboard_data(firstDate = '2023-06-24', Market = 'Six Swiss Exchange')
#> [1] "Page 0"
#> [1] "Page 0 has 10 observations."
#> [1] "Page 1"
#> [1] "Page 1 has 10 observations."
#> [1] "Page 2"
#> [1] "Page 2 has 10 observations."
#> [1] "Page 3"
#> [1] "Page 3 has 10 observations."
#> [1] "Page 4"
#> [1] "Page 4 has 10 observations."
#> [1] "Page 5"
#> [1] "Page 5 has 10 observations."
#> [1] "Page 6"
#> [1] "Page 6 has 10 observations."
#> [1] "Page 7"
#> [1] "Page 7 has 10 observations."
#> [1] "Page 8"
#> [1] "Page 8 has 10 observations."
#> [1] "Page 9"
#> [1] "Page 9 has 10 observations."
#> [1] "Page 10"
#> [1] "Page 10 has 10 observations."
#> [1] "Page 11"
#> [1] "Page 11 has 10 observations."
#> [1] "Page 12"
#> [1] "Page 12 has 10 observations."
#> [1] "Page 13"
#> [1] "Page 13 has 10 observations."
#> [1] "Page 14"
#> [1] "Page 14 has 10 observations."
#> [1] "Page 15"
#> [1] "Page 15 has 10 observations."
#> [1] "Page 16"
#> [1] "Page 16 has 10 observations."
#> [1] "Page 17"
#> [1] "Page 17 has 9 observations."
#> [1] "Page 18"
#> [1] "Page 18 has  observations."
head(data_final,1)
#>   messageNo         isin valorSymbol                                      title
#> 1    209038 LU0950674761      EMUSRI Mistrade Decision in EMUSRI / LU0950674761
#>   messageType broadcastDateTime                    security tradingSegment
#> 1    Mistrade       2.02402e+13 UBSETF MSCI EMU SRI EUR ACC            ETF
#>   priority markets products currency
#> 1   Normal    XSWX       FU      EUR
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                         newsText
#> 1 <div><p>In accordance with the rules of SIX Swiss Exchange, the following trade in <strong>'UBSETF MSCI EMU SRI EUR ACC'</strong> has been declared a mistrade and has therefore been cancelled: </p><table><thead><tr><th align="center" style="width: 75px">Trade Date</th><th align="center" style="width: 75px">Time</th><th align="center" style="width: 45px">Cur</th><th align="left" style="width: 75px">Size</th><th align="left" style="width: 75px">Price</th><th align="left" style="width: 75px">Trade Type</th><th align="left" style="width: 75px">Book Type</th><th align="left" style="width: 75px">Ref Exch</th></tr></thead><tbody><tr><td align="center">01.02.2024</td><td align="center">16:00:29</td><td align="center">EUR</td><td align="left">200</td><td align="left">22.6600</td><td align="left">OnExchange</td><td align="left">QuoteBook</td><td align="left"/></tr></tbody></table><p>Please find further information concerning mistrades in Directive 4: Market Control on our website.</p><p>Regards,<br/>Exchange Operations, SIX Swiss Exchange</p></div>
#>   newsTypeCode
#> 1           MI
data_final <- newsboard_data(firstDate = '2023-06-24', Products = 'Funds')
#> [1] "Page 0"
#> [1] "Page 0 has  observations."
head(data_final, 1)
#> data frame with 0 columns and 0 rows
if (FALSE) {
data_final <- newsboard_data(lastDate = '2021-07-01') # error when lastdate > firstdate
head(data_final, 1)
}

match
#> function (x, table, nomatch = NA_integer_, incomparables = NULL) 
#> .Internal(match(x, table, nomatch, incomparables))
#> <bytecode: 0x561a828ddc88>
#> <environment: namespace:base>