The Swiss Stock Exchange Market Data News via the SIX Group Newsboard

Returns the Swiss Stock Exchange Market Data News via the SIX Group Newsboard

Usage

newsboard_data(
  firstDate = Sys.Date() - 365 * 2,
  lastDate = Sys.Date(),
  messageType = c("Emergency Information", "Mistrade", "Suspension",
    "Trading Information"),
  Market = c("Six Swiss Exchange", "Structured Products",
    "XBTR - Bilateral Trading Plattform"),
  Products = c("Bonds", "Derivatives", "Equities", "Exchange Traded Products", "Funds")
)

Arguments

firstDate

a character string indicating the year-month-date of From date. The default is equivalent to firstDate = Sys.Date() - 365*2, the earliest date for which data is available on the Newsboard.

lastDate

a character string indicating the year-month-date of To date. The default is equivalent to `lastDate = Sys.Date()``, the last date in range for which data is available, but not greater than current date on the Newsboard.

messageType

a vector of character strings of the message type. This must be one or combination of either of the strings "Emergency Information", "Mistrade", "Suspension", "Trading Information".The default is equivalent to messageType = 'All'.

Market

a vector of character strings of the Market type. This must be one or combination of either of the strings "Six Swiss Exchange", "Structured Products","XBTR - Bilateral Trading Plattform". The default is equivalent to Market = 'All'.

Products

a vector of character strings of the products type. This must be one or combination of either of the strings "Bonds", "Derivatives", "Equities", "Exchange Traded Products", "Funds".The default is equivalent to Products = 'All'.

Value

Swiss Stock Exchange Market Data News via the SIX Group Newsboard

Details

The inputs must all be character strings, with the date format as %Y-%m-%d (as determined by as.Date()). The 'messageType', 'Market', 'Products', if not supplied, all character string options are used, which leads to the messageType = 'All', Market = 'All' and Products = 'All' defaults. Supplying individual parameter requires the use of complete string as opposed to the use of abbreviations.

References

Newsboard: Market activity at the Swiss Stock Exchange. SIX. https://www.six-group.com/en/products-services/the-swiss-stock-exchange/market-data/news-tools/newsboard.html#/

Author

Akhapwoya Obuchere

Examples

library(plyr)
library(jsonlite)

data_final <- newsboard_data(firstDate = '2023-06-24')
#> [1] "Page 0"
#> [1] "Page 0 has 10 observations."
#> [1] "Page 1"
#> [1] "Page 1 has 10 observations."
#> [1] "Page 2"
#> [1] "Page 2 has 10 observations."
#> [1] "Page 3"
#> [1] "Page 3 has 10 observations."
#> [1] "Page 4"
#> [1] "Page 4 has 10 observations."
#> [1] "Page 5"
#> [1] "Page 5 has 10 observations."
#> [1] "Page 6"
#> [1] "Page 6 has 10 observations."
#> [1] "Page 7"
#> [1] "Page 7 has 10 observations."
#> [1] "Page 8"
#> [1] "Page 8 has 10 observations."
#> [1] "Page 9"
#> [1] "Page 9 has 10 observations."
#> [1] "Page 10"
#> [1] "Page 10 has 10 observations."
#> [1] "Page 11"
#> [1] "Page 11 has 10 observations."
#> [1] "Page 12"
#> [1] "Page 12 has 10 observations."
#> [1] "Page 13"
#> [1] "Page 13 has 8 observations."
#> [1] "Page 14"
#> [1] "Page 14 has  observations."
head(data_final, 1)
#>   messageNo         isin valorSymbol                                   title
#> 1    208873 CH0344958688         E18 Mistrade Decision in E18 / CH0344958688
#>   messageType broadcastDateTime        security                  tradingSegment
#> 1    Mistrade      2.023101e+13 0.50 EIDG 18-32 Bonds - CHF Swiss Confederation
#>   priority markets products currency
#> 1   Normal    XSWX       BO      CHF
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                       newsText
#> 1 <div><p>In accordance with the rules of SIX Swiss Exchange, the following trade in <strong>'0.50 EIDG 18-32'</strong> has been declared a mistrade and has therefore been cancelled: </p><table><thead><tr><th align="center" style="width: 75px">Trade Date</th><th align="center" style="width: 75px">Time</th><th align="center" style="width: 45px">Cur</th><th align="left" style="width: 75px">Size</th><th align="left" style="width: 75px">Price</th><th align="left" style="width: 75px">Trade Type</th><th align="left" style="width: 75px">Book Type</th></tr></thead><tbody><tr><td align="center">11.10.2023</td><td align="center">11:21:55</td><td align="center">CHF</td><td align="left">10'000</td><td align="left">94.0000</td><td align="left">OnExchange</td><td align="left">Clob</td></tr></tbody></table><p>Please find further information concerning mistrades in Directive 4: Market Control on our website.</p><p>Regards,<br/>Exchange Operations, SIX Swiss Exchange</p></div>
#>   newsTypeCode
#> 1           MI
data_final <- newsboard_data(firstDate = '2023-06-24', messageType = 'Mistrade')
#> [1] "Page 0"
#> [1] "Page 0 has 10 observations."
#> [1] "Page 1"
#> [1] "Page 1 has 10 observations."
#> [1] "Page 2"
#> [1] "Page 2 has 10 observations."
#> [1] "Page 3"
#> [1] "Page 3 has 10 observations."
#> [1] "Page 4"
#> [1] "Page 4 has 10 observations."
#> [1] "Page 5"
#> [1] "Page 5 has 10 observations."
#> [1] "Page 6"
#> [1] "Page 6 has 10 observations."
#> [1] "Page 7"
#> [1] "Page 7 has 10 observations."
#> [1] "Page 8"
#> [1] "Page 8 has 10 observations."
#> [1] "Page 9"
#> [1] "Page 9 has 10 observations."
#> [1] "Page 10"
#> [1] "Page 10 has 10 observations."
#> [1] "Page 11"
#> [1] "Page 11 has 5 observations."
#> [1] "Page 12"
#> [1] "Page 12 has  observations."
head(data_final, 2)
#>   messageNo         isin valorSymbol                                      title
#> 1    208873 CH0344958688         E18    Mistrade Decision in E18 / CH0344958688
#> 2    208872 CH1272334934      KGIHJB Mistrade Decision in KGIHJB / CH1272334934
#>   messageType broadcastDateTime          security
#> 1    Mistrade      2.023101e+13   0.50 EIDG 18-32
#> 2    Mistrade      2.023101e+13 KGIHJB JB P 12/23
#>                    tradingSegment priority markets products currency
#> 1 Bonds - CHF Swiss Confederation   Normal    XSWX       BO      CHF
#> 2             Structured Products   Normal    XQMH       DE      CHF
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                            newsText
#> 1      <div><p>In accordance with the rules of SIX Swiss Exchange, the following trade in <strong>'0.50 EIDG 18-32'</strong> has been declared a mistrade and has therefore been cancelled: </p><table><thead><tr><th align="center" style="width: 75px">Trade Date</th><th align="center" style="width: 75px">Time</th><th align="center" style="width: 45px">Cur</th><th align="left" style="width: 75px">Size</th><th align="left" style="width: 75px">Price</th><th align="left" style="width: 75px">Trade Type</th><th align="left" style="width: 75px">Book Type</th></tr></thead><tbody><tr><td align="center">11.10.2023</td><td align="center">11:21:55</td><td align="center">CHF</td><td align="left">10'000</td><td align="left">94.0000</td><td align="left">OnExchange</td><td align="left">Clob</td></tr></tbody></table><p>Please find further information concerning mistrades in Directive 4: Market Control on our website.</p><p>Regards,<br/>Exchange Operations, SIX Swiss Exchange</p></div>
#> 2 <div><p>In accordance with the rules of SIX Swiss Exchange, the following trade in <strong>'KGIHJB JB P 12/23'</strong> has been declared a mistrade and has therefore been cancelled: </p><table><thead><tr><th align="center" style="width: 75px">Trade Date</th><th align="center" style="width: 75px">Time</th><th align="center" style="width: 45px">Cur</th><th align="left" style="width: 75px">Size</th><th align="left" style="width: 75px">Price</th><th align="left" style="width: 75px">Trade Type</th><th align="left" style="width: 75px">Book Type</th></tr></thead><tbody><tr><td align="center">10.10.2023</td><td align="center">16:40:20</td><td align="center">CHF</td><td align="left">5'000</td><td align="left">0.7600</td><td align="left">OnExchange</td><td align="left">QuoteBook</td></tr></tbody></table><p>Please find further information concerning mistrades in Directive 4: Market Control on our website.</p><p>Regards,<br/>Exchange Operations, SIX Swiss Exchange</p></div>
#>   newsTypeCode
#> 1           MI
#> 2           MI
data_final <- newsboard_data(firstDate = '2023-06-24', Market = 'Six Swiss Exchange')
#> [1] "Page 0"
#> [1] "Page 0 has 10 observations."
#> [1] "Page 1"
#> [1] "Page 1 has 10 observations."
#> [1] "Page 2"
#> [1] "Page 2 has 10 observations."
#> [1] "Page 3"
#> [1] "Page 3 has 10 observations."
#> [1] "Page 4"
#> [1] "Page 4 has 10 observations."
#> [1] "Page 5"
#> [1] "Page 5 has 10 observations."
#> [1] "Page 6"
#> [1] "Page 6 has 10 observations."
#> [1] "Page 7"
#> [1] "Page 7 has 8 observations."
#> [1] "Page 8"
#> [1] "Page 8 has  observations."
head(data_final,1)
#>   messageNo         isin valorSymbol                                   title
#> 1    208873 CH0344958688         E18 Mistrade Decision in E18 / CH0344958688
#>   messageType broadcastDateTime        security                  tradingSegment
#> 1    Mistrade      2.023101e+13 0.50 EIDG 18-32 Bonds - CHF Swiss Confederation
#>   priority markets products currency
#> 1   Normal    XSWX       BO      CHF
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                       newsText
#> 1 <div><p>In accordance with the rules of SIX Swiss Exchange, the following trade in <strong>'0.50 EIDG 18-32'</strong> has been declared a mistrade and has therefore been cancelled: </p><table><thead><tr><th align="center" style="width: 75px">Trade Date</th><th align="center" style="width: 75px">Time</th><th align="center" style="width: 45px">Cur</th><th align="left" style="width: 75px">Size</th><th align="left" style="width: 75px">Price</th><th align="left" style="width: 75px">Trade Type</th><th align="left" style="width: 75px">Book Type</th></tr></thead><tbody><tr><td align="center">11.10.2023</td><td align="center">11:21:55</td><td align="center">CHF</td><td align="left">10'000</td><td align="left">94.0000</td><td align="left">OnExchange</td><td align="left">Clob</td></tr></tbody></table><p>Please find further information concerning mistrades in Directive 4: Market Control on our website.</p><p>Regards,<br/>Exchange Operations, SIX Swiss Exchange</p></div>
#>   newsTypeCode
#> 1           MI
data_final <- newsboard_data(firstDate = '2023-06-24', Products = 'Funds')
#> [1] "Page 0"
#> [1] "Page 0 has  observations."
head(data_final, 1)
#> data frame with 0 columns and 0 rows
if (FALSE) {
data_final <- newsboard_data(lastDate = '2021-07-01') # error when lastdate > firstdate
head(data_final, 1)
}

match
#> function (x, table, nomatch = NA_integer_, incomparables = NULL) 
#> .Internal(match(x, table, nomatch, incomparables))
#> <bytecode: 0x5590392df188>
#> <environment: namespace:base>